Restructure components, move FullContainer
